+++ src/Main.cs
@@ -281,6 +281,7 @@ public class Initialisation : BaseUnityPlugin
                 Stereoscope.Init();
                 Flakseed.Init();
                 Xelsior.Init();
+                Empath.Init();
 
                 Lazy.FinalizeGuns(); // Make sure encounter trackables are finalized so shoot styles properly display in the Ammonomicon